webapp/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
itsdangerous>=2.2.0
authlib>=1.3.2
python-dotenv>=1.0.1
python-multipart>=0.0.9
pdfplumber>=0.11.0
pandas>=2.2.0
openpyxl>=3.1.2
yfinance>=0.2.40
